requirements.txt
django-sass-processor
base64
uuid
django-htmlmin
